repo directory tree restructure, fix illegal hyphen in the name - rename to skfp
