+++ skfp/__init__.py
@@ -1,0 +1,8 @@
+from skfp.fingerprints.morgan import MorganFingerprint
+from skfp.fingerprints.atom_pair import AtomPairFingerprint
+from skfp.fingerprints.topological_torsion import TopologicalTorsionFingerprint
+from skfp.fingerprints.maccs_keys import MACCSKeysFingerprint
+from skfp.fingerprints.erg import ERGFingerprint
+from skfp.fingerprints.map4 import MAP4Fingerprint
+from skfp.fingerprints.mhfp import MHFP
+from skfp.fingerprints.e3fp import E3FP
