+++ skfp/fingerprints/erg.py
@@ -1,0 +1,55 @@
+from typing import Union
+
+import numpy as np
+import pandas as pd
+import scipy.sparse as spsparse
+
+from skfp.fingerprints.base import FingerprintTransformer
+
+
+class ERGFingerprint(FingerprintTransformer):
+    def __init__(
+        self,
+        atom_types: int = 0,
+        fuzz_increment: float = 0.3,
+        min_path: int = 1,
+        max_path: int = 15,
+        sparse: bool = False,
+        n_jobs: int = None,
+        verbose: int = 0,
+        random_state: int = 0,
+        count: bool = False,  # unused
+    ):
+        super().__init__(
+            n_jobs=n_jobs,
+            sparse=sparse,
+            verbose=verbose,
+            random_state=random_state,
+            count=count,
+        )
+        self.atom_types = atom_types
+        self.fuzz_increment = fuzz_increment
+        self.min_path = min_path
+        self.max_path = max_path
+
+    def _calculate_fingerprint(
+        self, X: Union[pd.DataFrame, np.ndarray, list[str]]
+    ) -> Union[np.ndarray, spsparse.csr_array]:
+        X = self._validate_input(X)
+        from rdkit.Chem.rdReducedGraphs import GetErGFingerprint
+
+        X = [
+            GetErGFingerprint(
+                x,
+                atomTypes=self.atom_types,
+                fuzzIncrement=self.fuzz_increment,
+                minPath=self.min_path,
+                maxPath=self.max_path,
+            )
+            for x in X
+        ]
+
+        if self.sparse:
+            return spsparse.csr_array(X)
+        else:
+            return np.array(X)
