+++ skfp/fingerprints/maccs_keys.py
@@ -1,0 +1,37 @@
+from typing import Union
+
+import numpy as np
+import pandas as pd
+import scipy.sparse as spsparse
+
+from skfp.fingerprints.base import FingerprintTransformer
+
+
+class MACCSKeysFingerprint(FingerprintTransformer):
+    def __init__(
+        self,
+        sparse: bool = False,
+        n_jobs: int = 1,
+        verbose: int = 0,
+        random_state: int = 0,
+        count: bool = False,  # unused
+    ):
+        super().__init__(
+            n_jobs=n_jobs,
+            sparse=sparse,
+            verbose=verbose,
+            random_state=random_state,
+            count=count,
+        )
+
+    def _calculate_fingerprint(
+        self, X: Union[pd.DataFrame, np.ndarray, list[str]]
+    ) -> Union[np.ndarray, spsparse.csr_array]:
+        X = self._validate_input(X)
+        from rdkit.Chem.rdMolDescriptors import GetMACCSKeysFingerprint
+
+        X = [GetMACCSKeysFingerprint(x) for x in X]
+        if self.sparse:
+            return spsparse.csr_array(X)
+        else:
+            return np.array(X)
